refactor: 메인 페이지에 heart와 tag 적용
- PublicWorkbook => Workbook으로 변경(추상화)
- 자잘한 스타일 변경(margin 줄이기)
- WorkbookList css 스타일 추가(grid 속성이 중복되어서 추출)
