+++ frontend/src/styles/index.ts
@@ -1,3 +1,4 @@
 export { default as Flex } from './flex';
 export { default as loadContent } from './skeleton';
 export { default as scrollBarStyle } from './scroll';
+export { default as WorkbookListStyle } from './workbookList';
