+++ frontend/src/types/index.ts
@@ -12,13 +12,13 @@ export interface WorkbookResponse {
   name: string;
   cardCount: number;
   opened: boolean;
+  heartCount: number;
   tags: TagResponse[];
 }
 
 export interface PublicWorkbookResponse
   extends Omit<WorkbookResponse, 'opened'> {
   author: string;
-  heartCount: number;
 }
 
 export interface QuizResponse {
